Increment decrement cart product quantity implemented

+++ Services/cart/src/main/java/com/ecom/cart/entity/Cart.java
@@ -21,7 +21,7 @@ public class Cart extends BaseEntity {
 
     private BigDecimal totalPrice;
 
-    private BigDecimal discount;
+    private int discount;
 
     private BigDecimal total;
